+++ tsdocs.config.js
@@ -1,0 +1,5 @@
+// eslint-disable-next-line no-undef
+module.exports = {
+    "customPages": "./.docs",
+    "tsconfig": "./.docs/tsconfig.json"
+}